requirements.txt
pandas==2.2.2
numpy==2.0.1
faker==26.0.0
python-dateutil==2.9.0.post0
sqlalchemy==2.0.32
psycopg2-binary==2.9.9
pymongo==4.8.0
pyspark==3.5.1
